doit: Split out cmd_otp
